refactor: simplify route metadata strategy
